[#75] UITextField extension 으로 빼기
